refactor: avoid loading `babel-loader` when running esbuild
This commit refactor the code to avoid loading babel-loader when not needed.
